loginCallback without redirect
