+++ client/src/App.tsx
@@ -5,13 +5,18 @@ import './App.css';
 import Header from './components/Header';
 import ShoppingPage from './components/pages/ShoppingPage';
 import ProductPage from './components/pages/ProductPage';
+import NaverIdLogin from './components/pages/LoginPage';
+import NaverIdLoginCallback from './components/pages/LoginCallback';
 
 function App() {
   return (
     <>
       <Header />
       <Switch>
-        <Route path={['/']} component={HomePage} exact />
+        {/* <Route path={['/']} component={NaverIdLogin} exact /> */}
+        <Route path="/" component={NaverIdLogin} exact />
+        <Route path="/loginSuccess" component={NaverIdLoginCallback} />
+        <Route path="/home" component={HomePage} />
         <Route path="/channels/:channelID" component={ShoppingPage} />
         <Route path="/products/:productID" component={ProductPage} />
         <Route
